requirements.txt
Adafruit-Blinka==8.16.1
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-motor==3.4.9
adafruit-circuitpython-pca9685==3.4.8
adafruit-circuitpython-register==1.9.15
adafruit-circuitpython-requests==1.13.0
adafruit-circuitpython-servokit==1.3.13
adafruit-circuitpython-typing==1.9.0
Adafruit-PlatformDetect==3.42.0
Adafruit-PureIO==1.1.10
Bluetin-Echo==0.2.0
pkg-resources==0.0.0
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
typing-extensions==4.5.0
